add the keyboard handle example

+++ bpwpapcode/Chapter03/events.py
@@ -15,7 +15,7 @@ while True:
        
     event = pygame.event.wait()        
     event_text.append(str(event))
-    event_text = event_text[-SCREEN_SIZE[1]/font_height:]
+    event_text = event_text[int(-SCREEN_SIZE[1]/font_height):]
         
     if event.type == QUIT:
         exit()
